- Fix crashing in SeqDB / Release mode with very long database lists.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@33776 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbalias.hpp
@@ -1106,10 +1106,10 @@ private:
     bool m_IsProtein;
     
     /// Number of sequences.
-    mutable int m_NumSeqs;
+    mutable Int8 m_NumSeqs;
     
     /// Number of OIDs.
-    mutable int m_NumOIDs;
+    mutable Int8 m_NumOIDs;
     
     /// Total length.
     mutable Int8 m_TotalLength;
